refactor: unify quiz session hooks and remove dead code (#33)
- Delete 10 dead/redundant files: QuizScreen, ChoiceQuizScreen, MixedQuizScreen
  components (never imported), MixedQuizContent component, and the three
  separate session hooks (useChoiceQuizSession, useMixedQuizSession) with
  their test files
- Rewrite useQuizSession.ts as a unified hook that accepts a QuizMode parameter
  and handles 'type', 'choice', and 'mixed' modes via a single implementation
  based on the most complete useMixedQuizSession logic
- Export selectModeForWord and MAX_CONSECUTIVE_SAME_MODE from useQuizSession
  (confidence heuristic + consecutive-mode cap preserved)
- Rewrite ActiveQuizView.tsx with one container component that calls the single
  unified hook; switches between TypeQuizContent and ChoiceQuizContent based
  on state.currentMode per question
- Update ChoiceQuizContent to accept UseQuizSessionResult (unified type) and
  rename lastCorrect -> lastChoiceCorrect to match unified state
- Merge all three test files into one useQuizSession.test.ts with type mode,
  choice mode, and mixed mode describe blocks plus selectModeForWord unit tests
- Clean up index.ts to export only QuizHub, useQuizSession, and its types

All 330 tests pass, TypeScript strict clean, production build passes.
